% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanr-package.r
\docType{package}
\name{meanr-package}
\alias{meanr-package}
\title{meanr: Basic Sentiment Analysis Scorer}
\description{
A popular technique in text analysis today is sentiment
analysis, or trying to determine the overall emotional
attitude of a piece of text (positive or negative).
We provide a new, basic implementation of a common
method for computing sentiment, whereby words are scored
as positive or negative according to a "dictionary", and
then an average of those scores for the document is produced.
The package uses the Hu and Liu sentiment dictionary for
assigning sentiment.
}
\author{
Drew Schmidt \email{wrathematics AT gmail.com}
}
\keyword{Package}
