% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.r
\name{score}
\alias{score}
\title{score}
\usage{
score(s)
}
\arguments{
\item{s}{A character vector (vector of strings).}
}
\description{
Computes the sentiment score (simple average) of each string
in a vector of strings (independently).
}
\details{
The function uses the Hu and Liu sentiment dictionary (same
as everybody else) available here:
https://www.cs.uic.edu/~liub/FBS/sentiment-analysis.html
}
\references{
Hu, M., & Liu, B. (2004). Mining opinion features in customer
reviews. National Conference on Artificial Intelligence.
}

